Added sass precompiler and moved scss files to individual folders
