+++ front-end/.gitignore
@@ -9,6 +9,9 @@
 # production
 /build
 
+# css
+src/**/*.css
+
 # misc
 .DS_Store
 .env.local
