+++ front-end/src/components/App.scss
@@ -1,0 +1,16 @@
+.navigation {
+    display: flex;
+    position: fixed;
+    height: 4em;
+    width: 100%;
+    bottom: 0;
+    left: 0;
+    padding-top: 6px;
+    justify-content: space-around;
+    background-color: rgb(235, 235, 235);
+
+    a {
+        display: block;
+        text-decoration: none;
+    }
+  }
