+++ front-end/src/components/navigation-icon/NavigationIcon.scss
@@ -1,0 +1,17 @@
+.navigationIcon {
+    text-align: center;
+    font-size: 12px;
+    font-family: Calibri;
+    font-weight: bold;
+    text-decoration: none;
+    color: rgb(151, 151, 151);
+    margin-bottom: 8px;
+
+    svg {
+        height: 24px;
+    }
+
+    .active & {
+        color: rgb(60, 143, 222);
+    }
+}
