+++ front-end/src/components/newsfeed-view/Backgrounds.scss
@@ -1,0 +1,12 @@
+.backgrounds {
+    height: 55em;
+    width: 100%;
+    position: absolute;
+    z-index: -1;
+    top: 0;
+    left: 0;
+    background-image: url("../../assets/images/bg1.jpg");
+    background-repeat: no-repeat;
+    background-size: cover;
+    background-attachment: fixed;
+}
