+++ front-end/src/components/newsfeed-view/Header.scss
@@ -1,0 +1,3 @@
+.header {
+    color: #ffffff;
+}
